need to remove double quotes from values loaded into hpds.
Current methodology for building value masks in the data model requires
quoted strings for it to be parsed.  If a value has a quoted string
custom parsing will be needed to ingest a value with double quotes into
the value mask object.  A better methodology would be to store the value
mask used in the dictionary in the same manner used to store values in
the columnMeta.javabin.  If that is done then this work is not needed.